requirements.txt
numpy~=2.3.2
pillow~=11.3.0
matplotlib~=3.10.5
scipy~=1.16.1
scikit-learn~=1.7.1
scikit-image~=0.25.2
pip~=25.1.1
distro~=1.9.0
PySocks~=1.7.1
charset-normalizer~=3.3.2
setuptools~=80.3.1
cffi~=1.17.0
threadpoolctl~=3.6.0
packaging~=24.1
platformdirs~=4.2.2
joblib~=1.5.2
imageio~=2.37.0
rawpy~=0.25.1
tifffile~=2025.8.28
lazy_loader~=0.4
networkx~=3.5
six~=1.17.0
python-dateutil~=2.9.0.post0
contourpy~=1.3.3
fonttools~=4.59.2
pyparsing~=3.2.3
cycler~=0.12.1
kiwisolver~=1.4.9
psutil~=7.1.0
lazy_loader~=0.4